requirements.txt
matplotlib==3.6.0
numpy==1.23.3
Pillow==9.3.0
